Fix(users): GET /:id/brawlers 로직 수정 및 user_battles table column 수정

+++ src/database/migrations/1700655045629-brawltree.ts
@@ -120,7 +120,7 @@ export class Brawltree1700655045629 implements MigrationInterface {
                                  \`mode_code\`           tinyint                                                      NOT NULL,
                                  \`match_type\`          tinyint                                                      NOT NULL,
                                  \`match_grade\`         tinyint                                                      NOT NULL,
-                                 \`duration\`            tinyint unsigned DEFAULT NULL,
+                                 \`duration\`            smallint unsigned DEFAULT NULL,
                                  \`game_rank\`           tinyint                                                               DEFAULT NULL,
                                  \`game_result\`         tinyint                                                      NOT NULL,
                                  \`trophy_change\`       tinyint                                                      NOT NULL,
